- Vector3D integration started - walk the object v1 (!)
git-svn-id: https://svn.assembla.com/svn/archilogic@41 e0e46c49-be69-4f5a-ad62-21024a331aea

+++ Source/FacadePattern/src/ch/archilogic/math/geom/Line.java
@@ -36,7 +36,7 @@ public class Line {
 		return true;
 	}
 
-	public Vector3D getPoint(float t) {
+	public Vector3D getPoint(double t) {
 		return new Vector3D(P.getX()+t*D.getX(), P.getY()+t*D.getY(), P.getZ()+t*D.getZ());
 	}
 }
